feat: create carrousel component
